testutil: create private netns on the fly
Forking and re-executing the test binary in a new network namespace
has a bunch of drawbacks:

* Forces root even if running a test that doesn't need it
* Makes stracing a test more difficult
* Doesn't isolate tests against each other

Instead, add two helpers NewNetNS and JoinNetNS. The first creates a
new netns on the fly via unshare. This requires some trickery with
the go runtime to ensure that the migrated thread doesn't run other,
unrelated goroutines. The second allows joining the temporary netns
from a goroutine.