cmd/tubectl: add bind command
Allow creating bindings via the CLI tool. The invocation is close to
inet-tool, except that we expect port as a separate argument:

    bind label protocol prefix port

label may be any non-empty string (that isn't too long). protocol may
be udp or tcp, and prefix can either be a plain IP address (in which
case we assume a full netmask) or a CIDR. port 0 is special and means
"any port".