dispatcher: move label into Binding
Binging currently doesn't include a label, since that's the way we store
data in the BPF map. I think it's better to change this, so that
a Binding always includes a label. For a start, RemoveBinding becomes
a bit safer: we only remove a binding if the label matches. It also
means that if we end up changing the internal representation of a
Binding (e.g. if we split by IP version) our external API doesn't change.
Finally, Bindings() can return sorted []*Binding, which makes
implementing `tubectl list` easier.