TUBE-9: Use binary.Size for calculating the encoded Go struct size
unsafe.Sizeof returns the size of Go struct together with any
potential padding. Padding might be needed to guarantee natural
alignment for all fields in an array of structs.

While we're interested in the size of all fields without any padding,
because that's how the struct will be encoded when passing the value
to bpf(MAP_UPDATE).