TUBE-9: Switch to 64-bit label IDs without reuse
Reusing label IDs can lead to accidental associations between Bindings
and Destinations. Destinations will not hold a usage count on a label
because we won't be able to release the count when a Destination gets
removed due to a socket life-time event, e.g. socket got closed. Hence
we could end up in a situation where a Destinations gets re-associated
with a Binding that got a reused label ID.

Instead we will be wasteful with 64-bit IDs as they will not wrap
even if we were allocating IDs with high frequency for hundreds of
years.