allocate destination instead of label IDs
Right now we allocate an ID for a label when a binding is created
or a socket is added. We can leak label IDs since a socket may be
removed from a sockmap or sockhash by the kernel, without the
control plane noticing.

To fix this, we change the way we allocate IDs. Instead of
accounting for labels, we track (L3, L4, label) tuples. Like for
labels, we keep a reference count for each tuple. However, only
bindings take a reference on an ID.

Sockets on the other hand are implicitly refcounted by being
present in the sockmap. This means that we can end up with
an ID allocation that has a Count of 0. In such a case we
need to do a lookup in the sockmap to determine whether an
ID is unused.

We use this logic when allocating a new ID, which means that
"leaked" IDs are reclaimed at the point of allocation.

A nice side effect of this change is that we can use a sockmap
instead of a sockhash, since we only need a 32-bit index. The
code for sockmap is simpler than sockhash and hopefully less
buggy. Having a simple key to look up sockets with also means
that we now only need to deal with growing the sockmap, and
not changing the key format (at least we can hope).