Close BPF maps and program FDs after creating/opening the Dispatcher
We are leaking the BPF program FD from the tubectl main routine.  After
opening the Collection and moving the Map/Program ownership from the
Collection to dispatcherObjects, we hand off the Maps to destinations &
Dispatcher. The Program, however, remains owned by dispatcherObjects, which
we never close.

Switch to RAII approach, where dispatcherObjects owns the Program/Maps as
long as it lives, and close the dispatcherObjects when we're done with it.

Pass Cloned Maps to destinations/Dispatcher, which means we'll dup the FD.