TUBE-6: allow concurrent Dispatcher instances
We enforce a single Dispatcher instance per network namespace. This
forces a daemon to continually create and destroy Dispatchers if it
doesn't want to lock out the CLI tool. This in turn means that we
can't just drop privileges at the end of daemon start up.

Rework the Dispatcher locking to allow multiple instances of Dispatcher
to exist at the same time, but only allow a single one to mutate state.
A daemon can create or load a dispatcher, and drop privileges afterwards.

Creating a dispatcher now uses a temporary directory, which is renamed
to the real state directory once initialization is complete. This
removes a race condition where LoadDispatcher can observe an empty
state directory from CreateDispatcher and therefore return an error.

Whenever the Dispatcher has to read or modify maps it takes an
exclusive blocking lock on the state directory. This prevents other
Dispatchers from making concurrent changes.