check tag of loaded dispatcher
When loading a pinned dispatcher we currently do not check that the
running dispatcher is identical to the dispatcher that tubular expects.
This can in theory lead to modifications to the dispatcher state that
are incompatible with the running program. To avoid this, check
that the program tag matches, which is the truncated sha1 of the
eBPF bytecode (with certain references blanked out).

In the future we can use this as a signal to trigger an upgrade of
dispatcher state.