add wrappers for dealing with Unix cmsg
Unix ancilliary messages are not too involved in Go, since there is
support in x/sys/unix. There is a pitfall however: we need to make
sure that an attacker can't flood us with file descriptors and
thereby trigger EMFILE. For this reason we only allow receiving zero
or one file descriptors for now.