TUBE-9: Return a duplicate file descriptor from os.NewFile mock
os.NewFile mock used for testing FD passing is currently returning an
os.File that wraps the same file descriptor as does syscall.Conn passed in
ExtraFds.

This leads to unintended interaction between the test and the app logic. If
app closes its os.File, the file descriptor that is wraps gets closed. So
even though the test might be holding a reference to a net.Listener, the
backing socket will go away.

Ran into when adding a check that registering a socket inserts it into the
sockmap. Sockets where disappearing from the map, because the last FD for
the socket was getting closed sooner than expected, i.e. before the test
cleanup.