TUBE-37: use unlimited memlock when invoked via sudo
BPF programs are accounted against RLIMIT_MEMLOCK of the current user.
This means that loading the dispatcher may fail depending on which other
BPF has been loaded by the same user. This is most often a problem when
run as root, such as when invoked via sudo.

Bump RLIMIT_MEMLOCK if the process has CAP_SYS_RESOURCE: this allows
a per-user limit in /etc/security/limits.conf to take precedence
as long as the binary is invoked without the capability, while making
sure that invoking tubectl via sudo works as well.