dispatcher: release destination when replacing a binding
If a binding moves between two labels via a call to AddBinding, we
need to increase the refcount on the new destination ID and release
the refcount on the old destination ID. Unfortunately, this is more
complicated than it sounds: we only have a mapping of
(label, domain, proto) -> id at this point, not one that goes the
other way round. Use the brute force method of a linear search of
the allocation hash table: this is slow, but we currently only
allow 512 destinations anyways. Additionally, exact bindings moving
between labels is probably a rare occurrence.