cmd/tubectl: introduce a custom Logger interface
The tubectl environment currently takes stdout, stderr and a logger
(that is an alias for stdout). During tests, we use bytes.Buffer
for stderr, but testing.TB for the logger. This means that tests
for a stable output don't actually see all output.

At the same time the standard logger has functions we shouldn't use,
e.g. log.Fatalf. It also makes it easy to log to the default logger
instead of the passed in one. Make a custom Logger interface, which
is inspired by go-log. Make stderr and stdout in env loggers, and
use an in-memory logger to catch all output in tests.