TUBE-38: dispatcher: add ReplaceBindings
Allow replacing the currently active bindings in one shot. The method
figures out which bindings have been added (or updated) and removed.
It first adds new bindings and then removes deleted ones.

There is a change we will misdirect traffic:

1. add IP:80 -> foo
2. add IP:*  -> bar

If (2) is added before (1) a packet that arrives at the wrong time
will be send to bar instead of foo. The inverse is can happen when
removing bindings. For now we punt on this problem, see TUBE-45.