testutil: fix network namespace set up
The current network setup for a new namespace is only half done.
We add addresses to lo, but do not add routes. This works for IPv4
due to some kernel magic but breaks for IPv6. As an additional
wrinkle, the kernel creates a non-local route for IPv6 addresses.
Also, adding an address to an interface doesn't mean we can bind
to the whole subject. It just means that the subnet is reachable
via that interface. The kernel gives us address not found errors
due to that.

Fix this by:
* Adding explicit local routes before we add addresses.
* Using "nodad noprefixroute" when adding addresses to avoid
  automatic IPv6 routes and to disable duplicate address detection.
* Setting ip_nonlocal_bind to 1. This has the same effect as IP_FREEBIND
  but is netns specific.