remove SocketCookie from Destination
Destination contains SocketCookie so that we can print the cookie when
registering a new socket. This ends up being awkward for several reasons:

* In newDestinationFromBinding we set SocketCookie to 0, which isn't
  correct. We simply don't know whether there is a cookie.
* destinations.List has to do an additional sockmap lookup to find
  the cookie in the common case where alloc.Count > 0, even though
  most callers of List don't care about the cookie.
* Destination and destinationKey are meant to follow Binding and
  bindingKey, in that the *Key is the BPF representation of Binding and
  Destination. destinationKey doesn't contain the socket cookie.

We can fix this by providing alternative means to get the cookie only
when it's really needed. When registering we can simply get the
cookie directly from the socket and for metrics we add a separate
map for socket cookies.