dispatcher: move destinations map definition into BPF
We currently have a separate code path to create the hash map that
stores destination allocations since that state is only ever
accessed by the control plane, not the data plane. This makes
writing an upgrade function that takes all state into account
error prone.

Instead, define the destinations hash map in the BPF code. This means
that bpf2go auto generates struct fields for it. Patch up the parts
of the map definition that are only known at run time.