dispatcher: support unloading corrupt dispatcher instance
Unloading a dispatcher will currently fail if the persisted state
doesn't match what the code expects. This could be due to a bug
corrupting dispatcher state, or a legitimate change to the state
format.

Make UnloadDispatcher a free function that only assumes the minimum
needed to safely remove a dispatcher. This ensures that unload, load
is a viable disaster recovery strategy.