TUBE-52: dispatcher: add UpgradeDispatcher
Add a function which updates the program of a loaded dispatcher.
This covers the straightforward case when we change the dispatcher
program but keep maps the same across versions.

Changing maps is more complicated and needs to be tackled later.
For now we just assert that the existing maps are compatible with
the new program. If this is not the case a full unload, load cycle
is needed.