dispatcher: simplify locking
In commit cf31e97 I made Dispatcher locking more fine grained, with
the rationale that we need this for persistent services. It turns out this
is incorrect. Since then, we've discovered that OpenDispatcher can
proceed completely without any privileges, we only have to run as the
correct user. This means that long running processes can simply open
and close dispatchers as needed while being completely unprivileged.

Simplify locking again, so that there is only ever a single active
Dispatcher instance. This makes upgrades of Dispatcher state a lot
safer and reduces complexity. The downside is that we might have
to frequently open and close Dispatchers, but right now that doesn't
seem like a problem.