TUBE-53: dispatcher: allow read-only access
Allow opening a dispatcher in a read-only mode. This is achieved by loading
the pinned maps with BPF_F_RDONLY, which causes the kernel to deny writes.
Read-only mode takes a shared lock on the state directory, so multiple read-only
dispatchers will co-exist with each other.

Unfortunately, read-only semantics are not implemented for BPF programs and
links. This means we have to open programs and links in read-write mode, which
is a problem. As a work around, we disable the compatibility check for
read-only dispatchers.