TUBE-53: dispatcher: give everyone read-only access
The dispatcher currently uses mode 0700 for the state directory and
0600 for pinned maps and programs. This means that only the owning
user can access the state, which is needlessly restrictive.

Change the default modes to 0775 and 0664 for directories and files
respectively. The group has read-write access to state, while everybody
else gets read-only access. This means that operations like tubectl
list can be run without any privileges, by any user. Administration
can be delegated to a group of users and doesn't require privileges.