ebpf: change the maximum number of bindings
We currently limit the number of bindings to 4096, which is too low
to add all individual rrDNS IPs. Increase the limit to one million:
since LPM enforces BPF_F_NO_PREALLOC we there is no direct cost to
a high limit. One million seems sufficiently high so that we don't
ever have to raise it again (fingers crossed).