cmd/tubectl: remove serve command
tubectl serve was meant as a daemon that accepts register commands
on behalf of unprivileged services. That model is quite problematic,
since it requires adding the notion of an ACL. Instead of supporting
unprivileged register we will allow privileged users to register
sockets via pidfd_getfd.

Remove the serve command and the helpers for reading SCM_RIGHTS
messages.