TUBE-13: cmd/tubectl: refuse registering sockets when in different namespace
It's currently possible to register sockets with a dispatcher that are
from a different namespace. The result is that connections to those socket
are basically dropped. It would be nice to compare the network namespace of
each socket with the namespace of the dispatcher, but as far as I can
tell that isn't possible. There is no way to retrieve the network namespace
of a socket.

Instead, refuse to register sockets when the namespace of the register
process doesn't match the namespace of the dispatcher. In practice this
should offer similar protection.