update run-tests.sh
Pull in an updated version of run-tests.sh from cilium/ebpf. This
version compiles test binaries on the host and then executes them
in the VM. This is a lot more efficient than compiling the tests
inside of the VM.

The canonical way to run tests is now make test.