cmd/tubectl: add register-pid command
We can use pidfd_getfd to provide a very convenient way to use
tubular with unmodified services. We scan a target process'
file descriptors for ones that match a particular protocol, IP
and port and then register the found socket.

There are some subtleties here: first we need to figure out
when to stop iterating fds. I copied Marek's approach, which
simply stops after 32 contiguous fds are found to be unused.
Since the fd allocation fills early gaps this seems somewhat
safe. Second, we need to skip reuseport sockets, like we do
when registering systemd sockets. Third, we check that the
network namespace of the target process matches the dispatchers'.