TUBE-81: Add reachability check
We want to provide metrics that indicate if a binding is reachable.
Add a new Collector type called Reachable and use that for all
reachability metrics.

For now only support TCP. The check is simply a net.DialTimeout to the
first IP in the prefix. And registers a gauge with the number of
unreachable bindings.