TUBE-97: fix buggy dispatcher locking
Dispatcher locking has undergone a couple of changes. First we locked
the state directory for the lifetime of the Dispatcher object. Then
we made locking more fine grained, taking and releasing the lock
for each method on the dispatcher. Finally, we went back to coarse
locking since this simplifies upgrades significantly.

Except that we didn't remove all the fine grained locking. Fix this
and assert that dispatcher state is locked before the dispatcher
is being torn down.