TUBE-95: Prevent DoS from local user
Giving read permission to a directory allows a user to obtain a file
descriptor to it. With the file descriptor its possible to acquire a
lock using flock(). Since we give anyone on the system read access to
the state directory, anyone in the system can take a lock on it. As
long as the lock is taken no changes to bindings, sockets, etc. can
be made.

To fix this we have to change the file permissions of the state
directory. The owner still has full access, while the group has
read-only access. Other users have no access. An administrator can
give read-only access to state by adding users to the right group.